Deployment and testing (#2)
Improve testing and deployment

Testing: The unit tests are of two styles, one using the kubebuilder local api-server, and the other using the controller-runtime fakes. The first is the one recommended in the kubebuilder documentation, and is more like an end-to-end integration test suite. It uses the ginkgo framework, and the tests take slightly longer to run. The controller-runtime setup is lighter-weight and more like unit testing. Unfortunately it doesn't permit faking non-namespaced objects like CRDs.

This PR adds the gomock library to allow mocking out the Installer interface (and not depend on helm)

Deployment: A Helm chart is introduced. The RBAC configurations are fixed. Resource requirements are increased to avoid OOM-killing.

Other: Deletion is now supported, using a finalizer on the WaveComponent CR. The operator now updates the CR to set the version of the operator in an annotation, and a SparkVersion property in the status