Metrics configuration changes
I decided to add a new field into the WaveEnvironment CRD that
specifies wether we should enable the metrics subsystem.

I'm not sure we want to go this route tho we should really dicuss it
further.

Along those lines I did create an Environment abstraction that we can
reuse in other parts of the codebase as well as in spotctl.
This interface provides GetConfiguration and SaveConfiguration
methods.