Environment abstraction introduced (#44)
Along this some changes have been done to the tide manager
the controller runtime client and kubernetes clients are now
constructed at the same time as the manager and stored as members in the
struct. This greatly simplifies client usage in the manger since we no
longer have to start by constructing new clients each time.