[BGD-802] fix certificate issuer specification
This fix resolves the issue in the case of a new wave installation.

However, it does not work correctly when applying an upgrade, because
helm incorrectly merges the two annotations without removing the
first one. It's not possible to define the values so that it will
work correctly for as both initial installation and as a single-step upgrade.

Resolving the issue as an upgrade requires two steps -- to first disable
the component (which will cause helm to remove the chart) and then
enable it again.

EXCEPT: there's a difference between the enterprise-gateway and
spark-history-server charts. For the history server, the annotation
values are _quoted_, which means that the empty value for the cluster-issuer
annotation is passed through to helm, which fails with a nil value
at installation.
The enterprise-gateway annotation values are _unquoted_, which means that an
empty value *works* for initial installation, but if it is *omitted*
in the upgrade, it causes the nil value failure at upgrade.
Therefore, this fix removes the cluster-issue value in the history
server, but leaves it empty for the enterprise-gateway